+++ sql/V017__add_signoff_changelog.sql
@@ -1,4 +1,8 @@
 ALTER TABLE signoff ADD modified TIMESTAMP WITH TIME ZONE DEFAULT (NOW())::TIMESTAMP(0) WITH TIME ZONE NOT NULL;
+ALTER TABLE signoff ALTER COLUMN created TYPE TIMESTAMP(3) WITH TIME ZONE;
+ALTER TABLE signoff ALTER COLUMN created SET DEFAULT NOW();
+ALTER TABLE archive DROP CONSTRAINT IF EXISTS uniq_caseid_signoffstepname_deliverabletype;
+CREATE UNIQUE INDEX uniq_created_caseid_signoffstepname_deliverabletype ON signoff (created, case_identifier, signoff_step_name, deliverable_type);
 
 CREATE TABLE signoff_changelog (
   id SERIAL PRIMARY KEY,