+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project attempts to adhere to Semantic Versioning.
 
 ## UNRELEASED
 
+### Changed
 
+* no longer delete signoffs (keep all and return most recent)
 
 ## 3.14.1: [2026-01-30]
 
@@ -16,9 +18,11 @@ and this project attempts to adhere to Semantic Versioning.
 
 ## 3.14.0: [2026-01-26]
 
+
 ### Fixed
 
 * added documentation for the case archive resume-archiving endpoint
+
 
 ### Changed