+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project attempts to adhere to Semantic Versioning.
 ### Changed
 
 * when an archive entity is deleted, so is the archive record and the changelog
+* no longer delete signoffs (keep all and return most recent)
 
 ## 3.14.1: [2026-01-30]
 
@@ -18,9 +19,11 @@ and this project attempts to adhere to Semantic Versioning.
 
 ## 3.14.0: [2026-01-26]
 
+
 ### Fixed
 
 * added documentation for the case archive resume-archiving endpoint
+
 
 ### Changed